:construction_worker: Replacing relative for absolute URLs in the package description is done by altering the README file as a pre-build step. As a result,  doesn't require any external dependencies.
