+++ build.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+python build.py
+python setup.py sdist
